+++ examples/messaging_recvonly_sample/run.py
@@ -87,6 +87,7 @@ def install_deps(
             "macos_arm64",
             "ubuntu-22.04_x86_64",
             "ubuntu-24.04_x86_64",
+            "ubuntu-26.04_x86_64",
             "ubuntu-22.04_armv8",
             "ubuntu-24.04_armv8",
             "ubuntu-26.04_armv8",
@@ -145,6 +146,7 @@ def install_deps(
             "raspberry-pi-os_armv8",
             "ubuntu-22.04_x86_64",
             "ubuntu-24.04_x86_64",
+            "ubuntu-26.04_x86_64",
             "ubuntu-22.04_armv8",
             "ubuntu-24.04_armv8",
             "ubuntu-26.04_armv8",
@@ -179,6 +181,7 @@ AVAILABLE_TARGETS = [
     "macos_arm64",
     "ubuntu-22.04_x86_64",
     "ubuntu-24.04_x86_64",
+    "ubuntu-26.04_x86_64",
     "ubuntu-22.04_armv8",
     "ubuntu-24.04_armv8",
     "ubuntu-26.04_armv8",
@@ -244,7 +247,7 @@ def _build(args):
                 f"-DCMAKE_CXX_COMPILER={os.path.join(webrtc_info.clang_dir, 'bin', 'clang++')}",
                 f"-DLIBCXX_INCLUDE_DIR={cmake_path(os.path.join(webrtc_info.libcxx_dir, 'include'))}",
             ]
-        elif platform in ("ubuntu-22.04_x86_64", "ubuntu-24.04_x86_64"):
+        elif platform in ("ubuntu-22.04_x86_64", "ubuntu-24.04_x86_64", "ubuntu-26.04_x86_64"):
             cmake_args += [
                 f"-DCMAKE_C_COMPILER={os.path.join(webrtc_info.clang_dir, 'bin', 'clang')}",
                 f"-DCMAKE_CXX_COMPILER={os.path.join(webrtc_info.clang_dir, 'bin', 'clang++')}",